Feat/kriti Games APIs (#60)
* feat: Rule Schema and private APIs added

* feat: modularised codes

* feat:validation added

* ref: modularised codes

* feat: Validation added and insertion query added

* feat: validation added

* feat:kriti-updated the response

* feat: response JSON is updated

* feat: enhance rule validation and update database schema

* refactor: remove unnecessary console logs for cleaner code

* refactor: improve logging messages and clean up unused imports in game services

* feat: implement structured logging with Winston and add log directory

* feat: refactor logging system to support folder-specific loggers and improve directory management

* refactor: remove unused error transport imports for cleaner logger code

* update

* Update it to add DailyRotateFile that keep track of the errors based on dates

* Refact: Change all the console to logger

* feat: APIs for header added (wallet, notifications, favourite-Games)

* fix: Fix the Error Format for detailed Error Log

* Final: Logger Implementation Completed

* Final: Logger Implementation Completed

* feat: created games schema to store the details of each games

* feat: created games schema to store the details of each games

* feat: created games schema to store the details of each games

* Fix: Merging Process

* Fix: Merging Process

* updates made

* test/Kriti-checked APIs

* test/Kriti-checked APIs

* feat/Games List APIs added

* feat/Casino Games APIs added

---------
