+++ package.json
@@ -24,7 +24,7 @@
 		"bcryptjs": "^2.4.3",
 		"cors": "^2.8.5",
 		"dotenv": "^16.4.7",
-		"drizzle-orm": "^0.38.3",
+		"drizzle-orm": "^0.38.4",
 		"express": "^4.21.2",
 		"express-mysql-session": "^3.0.3",
 		"express-session": "^1.18.1",
