+++ src/controller/gameController/index.js
@@ -1,0 +1,16 @@
+import { getGameById } from "./getGamebyId.js";
+import { getGames } from "./getGames.js";
+import { getCurrentGame } from "./getCurrentGame.js";
+import { getGameHistory } from "./getGameHistory.js";
+import { placeBet, getValidBetOptions } from "./placeBet.js";
+import { getGamesByCategory } from "./getGamesByCategory.js";
+
+export {
+  getGameById,
+  getGames,
+  placeBet,
+  getValidBetOptions,
+  getGameHistory,
+  getCurrentGame,
+  getGamesByCategory,
+};
