+++ src/controller/notificationController/index.js
@@ -1,0 +1,3 @@
+import { addNotification } from "../notificationController/addNotification.js";
+import { getNotification } from "../notificationController/getNotification.js";
+export { addNotification, getNotification };
