+++ src/data/gamesDataByCategory.js
@@ -1,0 +1,57 @@
+// gamesData
+export const gamesDataByCategory = [
+  {
+    name: "TeenPatti T-20",
+    description: "Fast-paced card game",
+    category_id: 1,
+    thumbnail: "url_of_thumbnail",
+  },
+  {
+    name: "Andar Bahar 2",
+    description: "Classic Andar Bahar game",
+    category_id: 1,
+    thumbnail: "url_of_thumbnail",
+  },
+  {
+    name: "Lucky 7-B",
+    description: "Luck-based dice game",
+    category_id: 1,
+    thumbnail: "url_of_thumbnail",
+  },
+  {
+    name: "20-20 Dragon Tiger",
+    description: "Dragon Tiger game",
+    category_id: 1,
+    thumbnail: "url_of_thumbnail",
+  },
+  {
+    name: "Andar Bahar",
+    description: "Classic Andar Bahar game",
+    category_id: 1,
+    thumbnail: "url_of_thumbnail",
+  },
+  {
+    name: "TeenPatti T1-Day",
+    description: "Day version of TeenPatti T-20",
+    category_id: 1,
+    thumbnail: "url_of_thumbnail",
+  },
+  {
+    name: "20-20 D T L",
+    description: "Dragon Tiger Luck game",
+    category_id: 1,
+    thumbnail: "url_of_thumbnail",
+  },
+  {
+    name: "Lucky 7-A",
+    description: "Another version of Lucky 7",
+    category_id: 1,
+    thumbnail: "url_of_thumbnail",
+  },
+  {
+    name: "20-20 Dragon Tiger 2",
+    description: "Second version of Dragon Tiger",
+    category_id: 1,
+    thumbnail: "url_of_thumbnail",
+  },
+];
