+++ src/data/gamesListData.js
@@ -1,0 +1,63 @@
+// gamesListData.js
+export const gamesListData = [
+  {
+    name: "Indian Casino",
+    thumbnail: "url",
+    description: "Popular Indian casino games",
+  },
+  {
+    name: "Royal Casino",
+    thumbnail: "url",
+    description: "Royal-themed casino games",
+  },
+  {
+    name: "Aviator Casino",
+    thumbnail: "url",
+    description: "High-flying casino games",
+  },
+  {
+    name: "Game Dice",
+    thumbnail: "url",
+    description: "Classic dice games",
+  },
+  {
+    name: "Game Goal",
+    thumbnail: "url",
+    description: "Goal-oriented games",
+  },
+  {
+    name: "Det Casino",
+    thumbnail: "url",
+    description: "Detailed casino experience",
+  },
+  {
+    name: "Ezugi Casino",
+    thumbnail: "url",
+    description: "Live dealer casino games",
+  },
+  {
+    name: "Game Keno",
+    thumbnail: "url",
+    description: "Lottery-style game",
+  },
+  {
+    name: "Pragmatic Play",
+    thumbnail: "url",
+    description: "Innovative casino games",
+  },
+  {
+    name: "Game Hi-Lo",
+    thumbnail: "url",
+    description: "High-Low guessing game",
+  },
+  {
+    name: "Super Spade",
+    thumbnail: "url",
+    description: "Spade-themed games",
+  },
+  {
+    name: "Game Roulette",
+    thumbnail: "url",
+    description: "Classic roulette game",
+  },
+];
