+++ src/routes/publicApiRoute.js
@@ -16,4 +16,5 @@ router.post("/register", registerUser);
 router.post("/login", loginUser); // http://localhost:4320/api/login
 router.post("/logout", logoutUser);
 
+
 export default router;
